% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_nonendemics.R
\name{any_nonendemics}
\alias{any_nonendemics}
\title{Determine if the island community (DAISIE data) contains any non-endemic
lineages}
\usage{
any_nonendemics(daisie_data_list)
}
\arguments{
\item{daisie_data_list}{A single DAISIE data list (i.e. single replicate).}
}
\value{
A single logical.
}
\description{
Determine if the island community (DAISIE data) contains any non-endemic
lineages
}
