% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_param_diffs.R
\name{calc_param_diffs}
\alias{calc_param_diffs}
\title{Calculates the differences in parameter estimates (cladogenesis, extinction,
carrying capacity, colonisation, and anagenesis) from the DAISIE maximum
likelihood model fitted to the ideal and empirical data sets simulated from
`sim_island_with_mainland`.}
\usage{
calc_param_diffs(ml, param_set)
}
\arguments{
\item{param_set}{Numeric for the index of the parameter set from the
parameter space}
}
\value{
A list of five numeric vectors
}
\description{
Calculates the differences in parameter estimates (cladogenesis, extinction,
carrying capacity, colonisation, and anagenesis) from the DAISIE maximum
likelihood model fitted to the ideal and empirical data sets simulated from
`sim_island_with_mainland`.
}
\author{
Joshua W. Lambert
}
