% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for function arguments in the ContinentalTesting package}
\usage{
default_params_doc(
  signif,
  scientific,
  breaks,
  param_set,
  data_folder_path,
  output_file_path,
  transform,
  ml,
  total_time,
  m,
  island_pars,
  nonoceanic_pars,
  replicates,
  seed,
  verbose,
  sim_max_age,
  daisie_data_list
)
}
\arguments{
\item{signif}{A numeric specifying how many significant figures the axes
labels have when plotting}

\item{scientific}{A boolean determining whether the axis labels will be
converted to scientific notation}

\item{breaks}{A vector of numerics}

\item{param_set}{Numeric for the index of the parameter set from the
parameter space}

\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{transform}{Either `NULL` or `"ihs"` to specify no transformation of
the data or an inverse hyperbolic sine (ihs) transformation}

\item{ml}{A list of DAISIE data lists [DAISIE::DAISIE_ML_CS()] to calculate
the difference in parameter estimates}

\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{m}{Numeric defining the size of mainland pool.}

\item{island_pars}{A numeric vector containing the parameters for the island:
\itemize{
  \item{`island_pars[1]`: lambda^c (cladogenesis rate)}
  \item{`island_pars[2]`: mu (extinction rate)}
  \item{`island_pars[3]`: K (carrying capacity), set K=Inf for
  diversity independence.}
  \item{`island_pars[4]`: gamma (immigration rate)}
  \item{`island_pars[5]`: lambda^a (anagenesis rate)}
  }}

\item{nonoceanic_pars}{A vector of two numerics with the probability of
initial presence and the probability of a vicariant species being nonendemic
respectively}

\item{replicates}{Integer specifying number of island replicates to be
simulated.}

\item{seed}{Integer specifying the seed to set before simulating data}

\item{verbose}{Logical, determining if progress output should be printed
during the simulation}

\item{sim_max_age}{Logical, whether to augment the simulated DAISIE data to
convert vicariant species to max age (stacs)}

\item{daisie_data_list}{A single DAISIE data list (i.e. single replicate).}
}
\value{
Nothing
}
\description{
Documentation for function arguments in the ContinentalTesting package
}
\author{
Joshua W. Lambert
}
