% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_continental_data.R
\name{format_continental_data}
\alias{format_continental_data}
\title{Format the continental data simulated with DAISIE::DAISIE_sim_cr()}
\usage{
format_continental_data(sim)
}
\arguments{
\item{sim}{Single DAISIE data list (1 replicate), output by
[DAISIE::DAISIE_sim_cr()].}
}
\value{
DAISIE data list.
}
\description{
Changes the endemicity status (stac) of the simulated DAISIE data
in order to be analysed by the continental DAISIE inference model.
}
\keyword{internal}
