% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vicariance_loglik_diff.R
\name{plot_vicariance_loglik_diff}
\alias{plot_vicariance_loglik_diff}
\title{Plots a the difference in estimated maxmimum log likelihood between the
optimisation algorithms (\code{optimmethod}) for each parameter set}
\usage{
plot_vicariance_loglik_diff(
  data_folder_path,
  output_file_path,
  prob_init_pres_diff_threshold = NULL,
  x_axis = c("total_time", "prob_init_pres_diff")
)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{prob_init_pres_diff_threshold}{A \code{numeric} or \code{NULL} (default) for
whether to plot only those points that have a difference in probability of
initial presence (\code{prob_init_pres}) above a certain threshold (\verb{[0,1]}).}

\item{x_axis}{A \code{character} string of either \code{"total_time"} or
\code{"prob_init_pres_diff"} which determines which variable is used to plot
on the x_axis of the vicariance_loglik_diff plot.}
}
\value{
Void (saves plot)
}
\description{
Plots a the difference in estimated maxmimum log likelihood between the
optimisation algorithms (\code{optimmethod}) for each parameter set
}
