% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vicariance_scatter.R
\name{plot_vicariance_scatter}
\alias{plot_vicariance_scatter}
\title{Plots a the estimated probability of initial presence across vicariance
simulation scenarios}
\usage{
plot_vicariance_scatter(data_folder_path, output_file_path, parameter)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{A \code{character} string with either: \code{"prob_init_pres"} or
\code{"loglik"}.}
}
\value{
Void (saves plot)
}
\description{
Creates 2x2 facetted plot with the estimated vicariance parameter
(\code{prob_init_pres}) across optimisation methods, and \code{epss} values (i.e.
the difference between island age and maximum colonisation time).
Differences in simulated cladogenesis and anagensis rate are also plotted.
}
