% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_daisie_results.R
\name{post_process_daisie_results}
\alias{post_process_daisie_results}
\title{Compiles and saves DAISIE inference model output across the parameter space}
\usage{
post_process_daisie_results(data_folder_path, output_file_path)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}
}
\value{
Invisibly returns \code{"Finsished"} (used for saving file)
}
\description{
Compiles and saves DAISIE inference model output across the parameter space
}
