% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.continental_test}
\alias{print.continental_test}
\title{Print method for continental testing}
\usage{
\method{print}{continental_test}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns and prints object.
}
\description{
Print method for continental testing
}
