% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loglik_dist.R
\name{plot_loglik_dist}
\alias{plot_loglik_dist}
\title{Plots a grid of violin plots of the distribution of loglik values from the
DAISIE inference model across values of probability of initial presence}
\usage{
plot_loglik_dist(data_folder_path, output_file_path, parameter)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{A character string of either `"lambda_c"`, `"mu"`, `"K"`,
`"gamma"`, `"lambda_a"`, `"loglik"` to choose the parameter to plot}
}
\value{
Void (saves plot)
}
\description{
Plots a grid of violin plots of the distribution of loglik values from the
DAISIE inference model across values of probability of initial presence
}
