% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continental_island.R
\name{sim_continental_island}
\alias{sim_continental_island}
\title{Simulate and format continental island DAISIE data}
\usage{
sim_continental_island(
  total_time,
  m,
  island_pars,
  nonoceanic_pars,
  replicates,
  seed,
  verbose
)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{m}{Numeric defining the size of mainland pool.}

\item{island_pars}{A numeric vector containing the parameters for the island:
\itemize{
  \item{`island_pars[1]`: lambda^c (cladogenesis rate)}
  \item{`island_pars[2]`: mu (extinction rate)}
  \item{`island_pars[3]`: K (carrying capacity), set K=Inf for
  diversity independence.}
  \item{`island_pars[4]`: gamma (immigration rate)}
  \item{`island_pars[5]`: lambda^a (anagenesis rate)}
  }}

\item{replicates}{Integer specifying number of island replicates to be
simulated.}

\item{verbose}{Logical, determining if progress output should be printed
during the simulation.}
}
\value{
A list of DAISIE data lists ([DAISIE::DAISIE_sim_cr()]).
}
\description{
Simulate and format continental island DAISIE data
}
